pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gtrpgm-rule-engine"
version = "0.1.0"
description = "GTRPGM 게임의 규칙 위배 여부를 판정하는 엔진입니다."
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "fastapi-utils>=0.8.0",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "ruff>=0.14.11",
    "typing-inspect>=0.9.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools]
package-dir = { "" = "src" }
